Test codegen (#35)
Added functionality to test equivelance of the outputed kernels from
Exo. This has become important as we depend on more and more
user-defined operations that change their implementation continously.
This gives us confidence that outputed kernels remained unchanged and by
extension their performance.

I have currently enabled this for AVX2 github workflows and not for
Neon. This is because the focus for the next month will be on that
backend and there is not much point to maintaining those tests for now.